backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
langchain==0.1.0
langgraph==0.0.10
chromadb==0.4.22
neo4j==5.16.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.3
pypdf==3.17.4
tiktoken==0.5.2
sentence-transformers==2.2.2
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.26.0
